+++ packages/gleequore/tsconfig.json
@@ -7,11 +7,10 @@
     "esModuleInterop": true,
     "moduleResolution": "nodenext",
     "module": "NodeNext",
-    "rootDir": "./src"
+    "rootDir": "./src",
+    "declaration": true
   },
-  "include": [
-    "./src/**/*"
-  ],
+  "include": ["./src/**/*"],
   "exclude": [
     "node_modules",
     "lib",
